Adds validation on script name before deletion of backing file.

+++ dashboard/app/controllers/scripts_controller.rb
@@ -33,6 +33,13 @@ class ScriptsController < ApplicationController
   end
 
   def destroy
+    # Though script.name is restricted to containing lower case characters,
+    # dashes, and numbers only, we do these security checks for safety.
+    if (@script.name.start_with? '../') ||
+      (@script.name.include? '/../')
+      raise ArgumentError, "evil script name (#{@script.name})"
+    end
+
     @script.destroy
     filename = "config/scripts/#{@script.name}.script"
     File.delete(filename) if File.exist?(filename)
